requirements.txt
asgiref==1.1.2
attrs==17.2.0
autobahn==17.7.1
Automat==0.6.0
channels==1.1.6
constantly==15.1.0
daphne==1.3.0
dj-database-url==0.4.2
Django==1.11.4
djangorestframework==3.6.3
hyperlink==17.3.0
incremental==17.5.0
pytz==2017.2
six==1.10.0
Twisted==17.5.0
txaio==2.8.1
whitenoise==3.3.0
zope.interface==4.4.2
